sum of multiples of 3 or 5

+++ 0x02-functions_nested_loops/101-natural.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	long long int max = 1024;
+	int max = 1024;
 	int sum;
 	int n;
